+++ packages/terra-clinical-item-display/src/terra-dev-site/test/clinical-item-display/display/IconItemDisplay.test.jsx
@@ -2,15 +2,12 @@ import React from 'react';
 import IconSuccess from 'terra-icon/lib/icon/IconSuccess';
 
 import ItemDisplay from '../../../../ItemDisplay';
-
-const testStyle = {
-  padding: '10px 0',
-};
+import styles from './styles.module.scss';
 
 const testText = 'Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text Wrapping Test Text';
 
 export default () => (
-  <div style={testStyle}>
+  <div className={styles.spacer}>
     <ItemDisplay text={`Center / default - ${testText}`} icon={<IconSuccess />} id="test-icon-center" />
     <br />
     <ItemDisplay text={`Top - ${testText}`} icon={<IconSuccess />} id="test-icon-top" iconAlignment="top" />